agent: Fix Open Thread as Markdown not working when another panel is focused
